+++ drizzle/migrations/0085_wave57_settlement_fixes.sql
@@ -195,7 +195,7 @@ DECLARE
   v_bout record; v_mb text; v_wd boolean; v_terminal_void boolean;
   v_leg record; v_outcome text;
   v_p record; v_open int; v_lost int; v_won int; v_void int;
-  v_combined numeric; v_payout int; v_new_balance int;
+  v_combined numeric; v_payout bigint; v_new_balance bigint;
 BEGIN
   SELECT status::text AS status, winner_id, fighter_a_id, fighter_b_id,
          method::text AS method, round_finished
@@ -261,7 +261,7 @@ BEGIN
         ELSE
           SELECT LEAST(1000, round(exp(sum(ln(decimal_odds)))::numeric, 2))
             INTO v_combined FROM parlay_leg WHERE parlay_id=v_p.id AND status='won';
-          v_payout := floor(v_p.stake_coins * v_combined)::int;
+          v_payout := floor(v_p.stake_coins * v_combined)::bigint;
         END IF;
         UPDATE parlay SET status='won', payout=v_payout, settled_at=NOW() WHERE id=v_p.id;
         UPDATE user_profile SET balance_coins=balance_coins+v_payout,
